Add `IPlatformInformation` component and consolidate version info
Add a new component that is repsonsible for capturing OS/runtime version
information.

On Windows we use the .NET BCL OSVersion property to get the OS version
and P/Invoke to `IsOS` in shlwapi.dll to determine if the OS is a client
or server SKU. We include an app.manifest file stating compatibility
with Windows 8.1 & 10 to ensure the OSVersion property actually returns
the real values.

https://docs.microsoft.com/en-gb/windows/desktop/SysInfo/targeting-your-application-at-windows-8-1

On macOS we read the information from the SystemVersion.plist (and
equivalent on server SKUs). Another option here would be to shell out to
`sw_vers` and parse the version number. We cannot however use `uname` as
that returns information about the Darwin kernel, not the 'marketing'
name of macOS.

Linux has no implementation for platform information yet.