msauth: return MSAL wrapped result rather than JWT AT
Change the MSAuth GetToken method signature to return a wrapper result
object rather than just the JWT access token. The result wrapper object
includes the account UPN and the wrap AT as returned by MSAL. We do not
need to inspect the claims of the AT (which might not even by a JWT in
some cases anyway) since we have the account info elsewhere.