azrepos: add non-PAT based auth mode to AzRepos provider
Add a new setting to the Azure Repos provider that instructs GCM to
return the Azure access token directly, rather than use that token to
generate a new Azure DevOps Personal Access Token.

At the moment the only indication as to what account a user wants to use
is via the userinfo part of the remote URL for visualstudio.com style
URLs. This must be manually set.

The default configuration is still to continue to use PATs, for now.